+++ AddInManager/View/Control/LogControl.xaml.cs
@@ -37,6 +37,18 @@ namespace RevitAddinManager.View.Control
             this.Unloaded += viewModel.UserControl_Unloaded;
 
         }
-        
+        private void RightClickCopyCmdExecuted(object sender, ExecutedRoutedEventArgs e)
+        {
+            StringBuilder sb = new StringBuilder();
+            foreach (LogMessageString dd in listBox_LogMessages.SelectedItems)
+            {
+                sb.AppendLine(dd.Message);
+            }
+            Clipboard.SetText(sb.ToString());
+        }
+        private void RightClickCopyCmdCanExecute(object sender, CanExecuteRoutedEventArgs e)
+        {
+            e.CanExecute = listBox_LogMessages.SelectedItem != null;
+        }
     }
 }